refactor(cli): migrate stac tooling onto discovery
Move stac's refresher/writer/curated_ids/prober (with the asset-schema helpers)
to earthlens.stac.cli; add curated_collection_ids to the public toolkit (openeo/
eumetsat keep the core copy until they migrate). Relocate stac's asset-schema
unit tests to earthlens-imagery and repoint the core refresh_one/audit_one
orchestration tests' get_json seam to stac.cli. Shrink the guard allow-list.
